# backend/cmd/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Petplace API Description",
        "title": "Petplace API Version1",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
...
        "/client/again/{cage_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check Book Hotel Service Again",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Check Book Hotel Service Again",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by start_time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end_time",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/cleint/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "cancel or refund reservation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Cancel or Refund Reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund payload",
                        "name": "RefundPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RefundPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/client/booking": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "book hotel service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Book Hotel Service",
                "parameters": [
                    {
                        "description": "Booking payload",
                        "name": "BookingPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.BookingPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/client/review/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "review reservation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "review Reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review payload",
                        "name": "ReviewPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ReviewPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/client/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all hotel service user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Get all Hotel Service User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/client/{id}/{status}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get hotel service user by status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Get Hotel Service User by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/review/{profile_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get review hotel service by hotel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Get Review Hotel Service by Hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get hotel service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Get Hotel Service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accept or Reject a booking request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Accept or Reject a booking request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SelectStatus payload",
                        "name": "SelectStatusPayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SelectStatusPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/hotel/{id}/{status}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get hotel service hotel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HotelServices"
                ],
                "summary": "Get Hotel Service Hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.BookingPayload": {
            "type": "object",
            "required": [
                "animals",
                "cage_id",
                "card_detail",
                "client_id",
                "client_name",
                "end_time",
                "profile_id",
                "profile_name",
                "start_time"
            ],
            "properties": {
                "animals": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cage_id": {
                    "type": "integer"
                },
                "card_detail": {
                    "$ref": "#/definitions/types.CardPayload"
                },
                "client_id": {
                    "description": "payment information",
                    "type": "integer"
                },
                "client_name": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "integer"
                },
                "profile_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "types.CardPayload": {
            "type": "object",
            "properties": {
                "expiry": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "security_code": {
                    "type": "string"
                }
            }
        },
...
        "types.RefundPayload": {
            "type": "object",
            "required": [
                "client_id",
                "hotel_service_id",
                "paypal_email"
            ],
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "hotel_service_id": {
                    "type": "integer"
                },
                "paypal_email": {
                    "type": "string"
                }
            }
        },
        "types.ReviewPayload": {
            "type": "object",
            "required": [
                "hide_name",
                "hotel_service_id",
                "profile_id",
                "review_rate"
            ],
            "properties": {
                "hide_name": {
                    "type": "boolean"
                },
                "hotel_service_id": {
                    "type": "integer"
                },
                "profile_id": {
                    "type": "integer"
                },
                "review_detail": {
                    "type": "string"
                },
                "review_image": {
                    "type": "string"
                },
                "review_image_array": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "review_rate": {
                    "type": "number"
                }
            }
        },
        "types.SelectStatusPayload": {
            "type": "object",
            "required": [
                "hotel_service_id",
                "profile_id",
                "profile_name",
                "status"
            ],
            "properties": {
                "hotel_service_id": {
                    "type": "integer"
                },
                "profile_id": {
                    "type": "integer"
                },
                "profile_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}